+++ runtimes/eoapi/auth/eoapi/auth/__init__.py
@@ -1,0 +1,4 @@
+from .auth import OidcAuth  # noqa
+from .config import AuthSettings  # noqa
+
+__version__ = "0.1.0"
